requirements.txt
prometheus_client
kafka-python==2.0.2
psutil